Refactor DatabaseItemCRUDTest: Replace TestUtils with specific utility classes for improved organization and clarity in test data management
